#3645 Move to adempiere
#3645 Pharma: Import standard prices
