+++ de.metas.adempiere.adempiere/base/src/main/java-gen/de/metas/tax/model/I_C_VAT_SmallBusiness.java
@@ -1,0 +1,305 @@
+package de.metas.tax.model;
+
+
+/** Generated Interface for C_VAT_SmallBusiness
+ *  @author Adempiere (generated) 
+ */
+@SuppressWarnings("javadoc")
+public interface I_C_VAT_SmallBusiness 
+{
+
+    /** TableName=C_VAT_SmallBusiness */
+    public static final String Table_Name = "C_VAT_SmallBusiness";
+
+    /** AD_Table_ID=540112 */
+//    public static final int Table_ID = org.compiere.model.MTable.getTable_ID(Table_Name);
+
+//    org.compiere.util.KeyNamePair Model = new org.compiere.util.KeyNamePair(Table_ID, Table_Name);
+
+    /** AccessLevel = 3 - Client - Org
+     */
+//    java.math.BigDecimal accessLevel = java.math.BigDecimal.valueOf(3);
+
+    /** Load Meta Data */
+
+	/**
+	 * Get Mandant.
+	 * Mandant für diese Installation.
+	 *
+	 * <br>Type: TableDir
+	 * <br>Mandatory: true
+	 * <br>Virtual Column: false
+	 */
+	public int getAD_Client_ID();
+
+	public org.compiere.model.I_AD_Client getAD_Client();
+
+    /** Column definition for AD_Client_ID */
+    public static final org.adempiere.model.ModelColumn<I_C_VAT_SmallBusiness, org.compiere.model.I_AD_Client> COLUMN_AD_Client_ID = new org.adempiere.model.ModelColumn<I_C_VAT_SmallBusiness, org.compiere.model.I_AD_Client>(I_C_VAT_SmallBusiness.class, "AD_Client_ID", org.compiere.model.I_AD_Client.class);
+    /** Column name AD_Client_ID */
+    public static final String COLUMNNAME_AD_Client_ID = "AD_Client_ID";
+
+	/**
+	 * Set Sektion.
+	 * Organisatorische Einheit des Mandanten
+	 *
+	 * <br>Type: TableDir
+	 * <br>Mandatory: true
+	 * <br>Virtual Column: false
+	 */
+	public void setAD_Org_ID (int AD_Org_ID);
+
+	/**
+	 * Get Sektion.
+	 * Organisatorische Einheit des Mandanten
+	 *
+	 * <br>Type: TableDir
+	 * <br>Mandatory: true
+	 * <br>Virtual Column: false
+	 */
+	public int getAD_Org_ID();
+
+	public org.compiere.model.I_AD_Org getAD_Org();
+
+	public void setAD_Org(org.compiere.model.I_AD_Org AD_Org);
+
+    /** Column definition for AD_Org_ID */
+    public static final org.adempiere.model.ModelColumn<I_C_VAT_SmallBusiness, org.compiere.model.I_AD_Org> COLUMN_AD_Org_ID = new org.adempiere.model.ModelColumn<I_C_VAT_SmallBusiness, org.compiere.model.I_AD_Org>(I_C_VAT_SmallBusiness.class, "AD_Org_ID", org.compiere.model.I_AD_Org.class);
+    /** Column name AD_Org_ID */
+    public static final String COLUMNNAME_AD_Org_ID = "AD_Org_ID";
+
+	/**
+	 * Set Geschäftspartner.
+	 * Bezeichnet einen Geschäftspartner
+	 *
+	 * <br>Type: Search
+	 * <br>Mandatory: true
+	 * <br>Virtual Column: false
+	 */
+	public void setC_BPartner_ID (int C_BPartner_ID);
+
+	/**
+	 * Get Geschäftspartner.
+	 * Bezeichnet einen Geschäftspartner
+	 *
+	 * <br>Type: Search
+	 * <br>Mandatory: true
+	 * <br>Virtual Column: false
+	 */
+	public int getC_BPartner_ID();
+
+	public org.compiere.model.I_C_BPartner getC_BPartner();
+
+	public void setC_BPartner(org.compiere.model.I_C_BPartner C_BPartner);
+
+    /** Column definition for C_BPartner_ID */
+    public static final org.adempiere.model.ModelColumn<I_C_VAT_SmallBusiness, org.compiere.model.I_C_BPartner> COLUMN_C_BPartner_ID = new org.adempiere.model.ModelColumn<I_C_VAT_SmallBusiness, org.compiere.model.I_C_BPartner>(I_C_VAT_SmallBusiness.class, "C_BPartner_ID", org.compiere.model.I_C_BPartner.class);
+    /** Column name C_BPartner_ID */
+    public static final String COLUMNNAME_C_BPartner_ID = "C_BPartner_ID";
+
+	/**
+	 * Set Kleinunternehmer-Steuerbefreiung .
+	 *
+	 * <br>Type: ID
+	 * <br>Mandatory: true
+	 * <br>Virtual Column: false
+	 */
+	public void setC_VAT_SmallBusiness_ID (int C_VAT_SmallBusiness_ID);
+
+	/**
+	 * Get Kleinunternehmer-Steuerbefreiung .
+	 *
+	 * <br>Type: ID
+	 * <br>Mandatory: true
+	 * <br>Virtual Column: false
+	 */
+	public int getC_VAT_SmallBusiness_ID();
+
+    /** Column definition for C_VAT_SmallBusiness_ID */
+    public static final org.adempiere.model.ModelColumn<I_C_VAT_SmallBusiness, Object> COLUMN_C_VAT_SmallBusiness_ID = new org.adempiere.model.ModelColumn<I_C_VAT_SmallBusiness, Object>(I_C_VAT_SmallBusiness.class, "C_VAT_SmallBusiness_ID", null);
+    /** Column name C_VAT_SmallBusiness_ID */
+    public static final String COLUMNNAME_C_VAT_SmallBusiness_ID = "C_VAT_SmallBusiness_ID";
+
+	/**
+	 * Get Erstellt.
+	 * Datum, an dem dieser Eintrag erstellt wurde
+	 *
+	 * <br>Type: DateTime
+	 * <br>Mandatory: true
+	 * <br>Virtual Column: false
+	 */
+	public java.sql.Timestamp getCreated();
+
+    /** Column definition for Created */
+    public static final org.adempiere.model.ModelColumn<I_C_VAT_SmallBusiness, Object> COLUMN_Created = new org.adempiere.model.ModelColumn<I_C_VAT_SmallBusiness, Object>(I_C_VAT_SmallBusiness.class, "Created", null);
+    /** Column name Created */
+    public static final String COLUMNNAME_Created = "Created";
+
+	/**
+	 * Get Erstellt durch.
+	 * Nutzer, der diesen Eintrag erstellt hat
+	 *
+	 * <br>Type: Table
+	 * <br>Mandatory: true
+	 * <br>Virtual Column: false
+	 */
+	public int getCreatedBy();
+
+    /** Column definition for CreatedBy */
+    public static final org.adempiere.model.ModelColumn<I_C_VAT_SmallBusiness, org.compiere.model.I_AD_User> COLUMN_CreatedBy = new org.adempiere.model.ModelColumn<I_C_VAT_SmallBusiness, org.compiere.model.I_AD_User>(I_C_VAT_SmallBusiness.class, "CreatedBy", org.compiere.model.I_AD_User.class);
+    /** Column name CreatedBy */
+    public static final String COLUMNNAME_CreatedBy = "CreatedBy";
+
+	/**
+	 * Set Aktiv.
+	 * Der Eintrag ist im System aktiv
+	 *
+	 * <br>Type: YesNo
+	 * <br>Mandatory: true
+	 * <br>Virtual Column: false
+	 */
+	public void setIsActive (boolean IsActive);
+
+	/**
+	 * Get Aktiv.
+	 * Der Eintrag ist im System aktiv
+	 *
+	 * <br>Type: YesNo
+	 * <br>Mandatory: true
+	 * <br>Virtual Column: false
+	 */
+	public boolean isActive();
+
+    /** Column definition for IsActive */
+    public static final org.adempiere.model.ModelColumn<I_C_VAT_SmallBusiness, Object> COLUMN_IsActive = new org.adempiere.model.ModelColumn<I_C_VAT_SmallBusiness, Object>(I_C_VAT_SmallBusiness.class, "IsActive", null);
+    /** Column name IsActive */
+    public static final String COLUMNNAME_IsActive = "IsActive";
+
+	/**
+	 * Set Notiz.
+	 * Optional weitere Information für ein Dokument
+	 *
+	 * <br>Type: String
+	 * <br>Mandatory: false
+	 * <br>Virtual Column: false
+	 */
+	public void setNote (java.lang.String Note);
+
+	/**
+	 * Get Notiz.
+	 * Optional weitere Information für ein Dokument
+	 *
+	 * <br>Type: String
+	 * <br>Mandatory: false
+	 * <br>Virtual Column: false
+	 */
+	public java.lang.String getNote();
+
+    /** Column definition for Note */
+    public static final org.adempiere.model.ModelColumn<I_C_VAT_SmallBusiness, Object> COLUMN_Note = new org.adempiere.model.ModelColumn<I_C_VAT_SmallBusiness, Object>(I_C_VAT_SmallBusiness.class, "Note", null);
+    /** Column name Note */
+    public static final String COLUMNNAME_Note = "Note";
+
+	/**
+	 * Get Aktualisiert.
+	 * Datum, an dem dieser Eintrag aktualisiert wurde
+	 *
+	 * <br>Type: DateTime
+	 * <br>Mandatory: true
+	 * <br>Virtual Column: false
+	 */
+	public java.sql.Timestamp getUpdated();
+
+    /** Column definition for Updated */
+    public static final org.adempiere.model.ModelColumn<I_C_VAT_SmallBusiness, Object> COLUMN_Updated = new org.adempiere.model.ModelColumn<I_C_VAT_SmallBusiness, Object>(I_C_VAT_SmallBusiness.class, "Updated", null);
+    /** Column name Updated */
+    public static final String COLUMNNAME_Updated = "Updated";
+
+	/**
+	 * Get Aktualisiert durch.
+	 * Nutzer, der diesen Eintrag aktualisiert hat
+	 *
+	 * <br>Type: Table
+	 * <br>Mandatory: true
+	 * <br>Virtual Column: false
+	 */
+	public int getUpdatedBy();
+
+    /** Column definition for UpdatedBy */
+    public static final org.adempiere.model.ModelColumn<I_C_VAT_SmallBusiness, org.compiere.model.I_AD_User> COLUMN_UpdatedBy = new org.adempiere.model.ModelColumn<I_C_VAT_SmallBusiness, org.compiere.model.I_AD_User>(I_C_VAT_SmallBusiness.class, "UpdatedBy", org.compiere.model.I_AD_User.class);
+    /** Column name UpdatedBy */
+    public static final String COLUMNNAME_UpdatedBy = "UpdatedBy";
+
+	/**
+	 * Set Gültig ab.
+	 * Gültig ab inklusiv (erster Tag)
+	 *
+	 * <br>Type: Date
+	 * <br>Mandatory: true
+	 * <br>Virtual Column: false
+	 */
+	public void setValidFrom (java.sql.Timestamp ValidFrom);
+
+	/**
+	 * Get Gültig ab.
+	 * Gültig ab inklusiv (erster Tag)
+	 *
+	 * <br>Type: Date
+	 * <br>Mandatory: true
+	 * <br>Virtual Column: false
+	 */
+	public java.sql.Timestamp getValidFrom();
+
+    /** Column definition for ValidFrom */
+    public static final org.adempiere.model.ModelColumn<I_C_VAT_SmallBusiness, Object> COLUMN_ValidFrom = new org.adempiere.model.ModelColumn<I_C_VAT_SmallBusiness, Object>(I_C_VAT_SmallBusiness.class, "ValidFrom", null);
+    /** Column name ValidFrom */
+    public static final String COLUMNNAME_ValidFrom = "ValidFrom";
+
+	/**
+	 * Set Gültig bis.
+	 * Gültig bis inklusiv (letzter Tag)
+	 *
+	 * <br>Type: Date
+	 * <br>Mandatory: true
+	 * <br>Virtual Column: false
+	 */
+	public void setValidTo (java.sql.Timestamp ValidTo);
+
+	/**
+	 * Get Gültig bis.
+	 * Gültig bis inklusiv (letzter Tag)
+	 *
+	 * <br>Type: Date
+	 * <br>Mandatory: true
+	 * <br>Virtual Column: false
+	 */
+	public java.sql.Timestamp getValidTo();
+
+    /** Column definition for ValidTo */
+    public static final org.adempiere.model.ModelColumn<I_C_VAT_SmallBusiness, Object> COLUMN_ValidTo = new org.adempiere.model.ModelColumn<I_C_VAT_SmallBusiness, Object>(I_C_VAT_SmallBusiness.class, "ValidTo", null);
+    /** Column name ValidTo */
+    public static final String COLUMNNAME_ValidTo = "ValidTo";
+
+	/**
+	 * Set Umsatzsteuer-ID.
+	 *
+	 * <br>Type: String
+	 * <br>Mandatory: false
+	 * <br>Virtual Column: false
+	 */
+	public void setVATaxID (java.lang.String VATaxID);
+
+	/**
+	 * Get Umsatzsteuer-ID.
+	 *
+	 * <br>Type: String
+	 * <br>Mandatory: false
+	 * <br>Virtual Column: false
+	 */
+	public java.lang.String getVATaxID();
+
+    /** Column definition for VATaxID */
+    public static final org.adempiere.model.ModelColumn<I_C_VAT_SmallBusiness, Object> COLUMN_VATaxID = new org.adempiere.model.ModelColumn<I_C_VAT_SmallBusiness, Object>(I_C_VAT_SmallBusiness.class, "VATaxID", null);
+    /** Column name VATaxID */
+    public static final String COLUMNNAME_VATaxID = "VATaxID";
+}
