+++ de.metas.adempiere.adempiere/base/src/main/java-gen/de/metas/tax/model/X_C_VAT_SmallBusiness.java
@@ -1,19 +1,3 @@
-/******************************************************************************
- * Product: Adempiere ERP & CRM Smart Business Solution                       *
- * Copyright (C) 1999-2007 ComPiere, Inc. All Rights Reserved.                *
- * This program is free software, you can redistribute it and/or modify it    *
- * under the terms version 2 of the GNU General Public License as published   *
- * by the Free Software Foundation. This program is distributed in the hope   *
- * that it will be useful, but WITHOUT ANY WARRANTY, without even the implied *
- * warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.           *
- * See the GNU General Public License for more details.                       *
- * You should have received a copy of the GNU General Public License along    *
- * with this program, if not, write to the Free Software Foundation, Inc.,    *
- * 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA.                     *
- * For the text or an alternative of this public license, you may reach us    *
- * ComPiere, Inc., 2620 Augustine Dr. #245, Santa Clara, CA 95054, USA        *
- * or via info@compiere.org or http://www.compiere.org/license.html           *
- *****************************************************************************/
 /** Generated Model - DO NOT CHANGE */
 package de.metas.tax.model;
 
@@ -30,7 +14,7 @@ public class X_C_VAT_SmallBusiness extends org.compiere.model.PO implements I_C_
 	/**
 	 *
 	 */
-	private static final long serialVersionUID = -2131683775L;
+	private static final long serialVersionUID = -1161208034L;
 
     /** Standard Constructor */
     public X_C_VAT_SmallBusiness (Properties ctx, int C_VAT_SmallBusiness_ID, String trxName)
@@ -60,14 +44,6 @@ public class X_C_VAT_SmallBusiness extends org.compiere.model.PO implements I_C_
       return poi;
     }
 
-    @Override
-    public String toString()
-    {
-      StringBuffer sb = new StringBuffer ("X_C_VAT_SmallBusiness[")
-        .append(get_ID()).append("]");
-      return sb.toString();
-    }
-
 	@Override
 	public org.compiere.model.I_C_BPartner getC_BPartner() throws RuntimeException
 	{
@@ -88,9 +64,9 @@ public class X_C_VAT_SmallBusiness extends org.compiere.model.PO implements I_C_
 	public void setC_BPartner_ID (int C_BPartner_ID)
 	{
 		if (C_BPartner_ID < 1) 
-			set_Value (COLUMNNAME_C_BPartner_ID, null);
+			set_ValueNoCheck (COLUMNNAME_C_BPartner_ID, null);
 		else 
-			set_Value (COLUMNNAME_C_BPartner_ID, Integer.valueOf(C_BPartner_ID));
+			set_ValueNoCheck (COLUMNNAME_C_BPartner_ID, Integer.valueOf(C_BPartner_ID));
 	}
 
 	/** Get Geschäftspartner.
